Crate aoc_2017_day_16

Advent of Code 2017: Day 16: Permutation Promenade

https://adventofcode.com/2017/day/16

Input

A single line containing a comma-separated list of “dance steps”, where X and Y are positions in the range 0 to 15, and A and B are programmes in the range a to p.

sX rotates the line left by X

xX/Y exchanges positions X and Y

pA/B swaps programmes A and B

Part one

Starting with the programme line abcdefghijklmnop, apply the steps once.

Part two

Same as part one, but repeat the process one billion times.

Algorithm

We split the process in two, tracking position changes separately from programme swaps, which we refer to as a transformation. These transformations can then be combined giving an efficient way to execute many steps in O(n) time. To apply a transformation we first apply the position changes, then the programme swaps.

First we combine all of the steps into a single transformation. We then iterate over the bits in the number of times we want to apply the steps, starting with the lowest, updating the line if the bit is a one. To work out the transformation for the next bit, we combine the current transformation with itself.

Modules

input

logic

Logic for solving the puzzles

model

Functions

puzzle_metadata